Saml Config rename + actual saml config component
