+++ src/components/SamlConfiguration/index.jsx
@@ -1,280 +1,71 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import moment from 'moment';
-import { Collapsible, Icon } from '@edx/paragon';
-import classNames from 'classnames';
-import SamlProviderConfigForm from './SamlProviderConfigForm';
-import SamlProviderDataForm from './SamlProviderDataForm';
-import { camelCaseObject, snakeCaseFormData } from '../../utils';
+import { ValidationFormGroup, Input } from '@edx/paragon';
+import LMSApiService from '../../data/services/LmsApiService';
 import NewRelicService from '../../data/services/NewRelicService';
-import LmsApiService from '../../data/services/LmsApiService';
-import LoadingMessage from '../LoadingMessage';
-import ErrorPage from '../ErrorPage';
+
 
 class SamlConfiguration extends React.Component {
   state = {
-    providerConfig: undefined,
-    providerData: undefined,
-    error: undefined,
+    configs: [],
     loading: true,
-    deleteEnabled: false,
   };
 
   componentDidMount() {
-    Promise.allSettled([
-      LmsApiService.getProviderConfig(this.props.enterpriseId),
-      LmsApiService.getProviderData(this.props.enterpriseId),
-    ]).then((responses) => {
-      this.setState({
-        providerConfig: responses[0].status === 'fulfilled' ? responses[0].value.data.results[0] : undefined,
-        providerData: responses[1].status === 'fulfilled' ? responses[1].value.data.results[0] : undefined,
+    LMSApiService.fetchSamlConfigurations()
+      .then(response => this.setState({
+        configs: response.data.results,
         loading: false,
-      });
-    })
+      }))
       .catch((error) => {
         const errorMsg = error.message || error.response.status === 500 ?
           error.message : JSON.stringify(error.response.data);
         NewRelicService.logAPIErrorResponse(errorMsg);
         this.setState({
-          error,
           loading: false,
         });
       });
   }
 
-  handleErrors(error) {
-    const errorMsg = error.message || error.response?.status === 500 ?
-      error.message : JSON.stringify(error.response.data);
-    NewRelicService.logAPIErrorResponse(errorMsg);
-    return errorMsg;
-  }
-
-  /**
-   * Creates a new third party provider configuration, then updates this list with the response.
-   * Returns if there is an error.
-   * @param {FormData} formData
-   */
-  createProviderConfig = async (formData) => {
-    const transformedData = snakeCaseFormData(formData);
-    transformedData.append('name', this.props.enterpriseName);
-    transformedData.append('slug', this.props.enterpriseSlug);
-    transformedData.append('enterprise_customer_uuid', this.props.enterpriseId);
-    try {
-      const response = await LmsApiService.postNewProviderConfig(transformedData);
-      this.setState({ providerConfig: response.data });
-      return undefined;
-    } catch (error) {
-      return this.handleErrors(error);
-    }
-  }
-
-  /**
-   * Updates existing provider configuration, then updates this list with the response.
-   * Returns if there is an error.
-   * @param {FormData} formData
-   * @param {string} pid // provider ID generated and stored in database.
-   */
-  updateProviderConfig = async (formData, pid) => {
-    const transformedData = snakeCaseFormData(formData);
-    transformedData.append('name', this.props.enterpriseName);
-    transformedData.append('slug', this.props.enterpriseSlug);
-    transformedData.append('enterprise_customer_uuid', this.props.enterpriseId);
-    try {
-      const response = await LmsApiService.updateProviderConfig(transformedData, pid);
-      this.setState({ providerConfig: response.data });
-      return undefined;
-    } catch (error) {
-      return this.handleErrors(error);
-    }
-  }
-
-  /**
-   * Permanently deletes a provider configuration, then updates this list with the response.
-   * Returns if there is an error.
-   * @param {string} pid // provider ID generated and stored in database.
-   */
-  deleteProviderConfig = async (pid) => {
-    try {
-      await LmsApiService.deleteProviderConfig(pid, this.props.enterpriseId);
-      this.setState({ providerConfig: undefined });
-      return undefined;
-    } catch (error) {
-      return this.handleErrors(error);
-    }
-  }
-
-  /**
-   * Creates new third party provider data entry in db and refreshes list with the result.
-   * Returns if there is an error.
-   * @param {FormData} formData
-   */
-  createProviderData = async (formData) => {
-    formData.append('fetchedAt', moment().format('YYYY-MM-DDThh:mm:ss'));
-    formData.append('enterpriseCustomerUuid', this.props.enterpriseId);
-    const transformedData = snakeCaseFormData(formData);
-    try {
-      const response = await LmsApiService.createProviderData(transformedData);
-      this.setState({ providerData: response.data });
-      return undefined;
-    } catch (error) {
-      return this.handleErrors(error);
-    }
-  }
-
-  /**
-   * Deletes a provider data entry in db and refreshes list with the result.
-   * Returns if there is an error.
-   * @param {string} pdid // provider data id generated & stored in database.
-   */
-  deleteProviderData = async (pdid) => {
-    try {
-      await LmsApiService.deleteProviderData(pdid, this.props.enterpriseId);
-      this.setState({ providerData: undefined });
-      return undefined;
-    } catch (error) {
-      return this.handleErrors(error);
-    }
+  getConfigOptions() {
+    const { configs } = this.state;
+    const options = [];
+    options.push({ label: '-- choose a configuration --', value: '', hidden: true });
+    configs.forEach((object) => {
+      options.push({ label: object.slug, value: object.id });
+    });
+    return options;
   }
 
   render() {
-    const {
-      providerConfig, error, providerData, loading, deleteEnabled,
-    } = this.state;
-
-    if (loading) {
-      return <LoadingMessage className="overview" />;
-    }
-    if (error && error.response?.status !== 404) {
-      return (
-        <ErrorPage
-          status={error.response?.status}
-          message={error.message}
-        />
-      );
-    }
     return (
-      <main role="main">
-        <div>
-          <React.Fragment>
-            {providerConfig &&
-              <div
-                key={providerConfig.id}
-                className="mb-3"
-              >
-                <Collapsible
-                  styling="card"
-                  className="shadow"
-                  title={
-                    <div className="row justify-content-around flex-fill">
-                      <Icon
-                        className={classNames(
-                          'col-1',
-                          {
-                            'fa fa-check text-success-300': providerConfig.enabled,
-                            'fa fa-times text-danger-300': !providerConfig.enabled,
-                          },
-                        )}
-                      />
-                      <div className="col">
-                        <h3 className="h6">Entity ID:</h3>
-                        <p>{providerConfig.entity_id}</p>
-                      </div>
-                      <div className="col">
-                        <h3 className="h6">Metadata Source:</h3>
-                        <p>{providerConfig.metadata_source}</p>
-                      </div>
-                    </div>
-                      }
-                >
-                  <SamlProviderConfigForm
-                    config={camelCaseObject(providerConfig)}
-                    updateProviderConfig={this.updateProviderConfig}
-                    createProviderConfig={this.createProviderConfig}
-                    deleteProviderConfig={this.deleteProviderConfig}
-                    deleteEnabled={deleteEnabled}
-                  />
-                </Collapsible>
-              </div>
-            }
-            {providerData &&
-              <div
-                key={providerData.id}
-                className="mb-3"
-              >
-                <Collapsible
-                  styling="card"
-                  className="shadow"
-                  title={
-                    <div className="row justify-content-around flex-fill">
-                      <div className="col">
-                        <h3 className="h6">Fetched At:</h3>
-                        <p>{providerData.fetched_at}</p>
-                      </div>
-                      <div className="col">
-                        <h3 className="h6">Entity ID:</h3>
-                        <p>{providerData.entity_id}</p>
-                      </div>
-                      <div className="col">
-                        <h3 className="h6">SSO URL:</h3>
-                        <p>{providerData.sso_url}</p>
-                      </div>
-                    </div>
-                  }
-                >
-                  <SamlProviderDataForm
-                    pData={camelCaseObject(providerData)}
-                    deleteProviderData={this.deleteProviderData}
-                    deleteEnabled={deleteEnabled}
-                  />
+      <div className="col col-4">
+        <ValidationFormGroup
+          for="samlConfigId"
+          helpText="the edX certificates to use with your SAML provider."
+        >
+          <label htmlFor="samlConfigId">Saml Configuration</label>
 
-                </Collapsible>
-              </div>
-            }
-            {!providerConfig &&
-              <Collapsible
-                styling="basic"
-                title="Add a SAML Provider Configuration"
-                className="col justify-content-center align-items-center"
-              >
-                <div>
-                  <SamlProviderConfigForm
-                    createProviderConfig={this.createProviderConfig}
-                    deleteEnabled={deleteEnabled}
-                  />
-                </div>
-              </Collapsible>
-            }
-            {(providerConfig && !providerData) &&
-              <Collapsible
-                styling="basic"
-                title="Add SAML Provider Data"
-                className="col justify-content-center align-items-center"
-              >
-                <div>
-                  <SamlProviderDataForm
-                    entityId={providerConfig.entity_id}
-                    createProviderData={this.createProviderData}
-                    deleteEnabled={deleteEnabled}
-                  />
-                </div>
-              </Collapsible>
-            }
-          </React.Fragment>
-        </div>
-      </main>
+          <Input
+            type="select"
+            id="samlConfigId"
+            name="samlConfigId"
+            key={this.state.loading ? 'loaded' : 'loading'}
+            options={this.getConfigOptions()}
+            defaultValue={this.props.currentConfig}
+          />
+        </ValidationFormGroup>
+      </div>
     );
   }
 }
 
 SamlConfiguration.defaultProps = {
-  enterpriseSlug: null,
-  enterpriseName: null,
+  currentConfig: undefined,
 };
 
 SamlConfiguration.propTypes = {
-  enterpriseName: PropTypes.string,
-  enterpriseSlug: PropTypes.string,
-  enterpriseId: PropTypes.string.isRequired,
+  currentConfig: PropTypes.number,
 };
 
 export default SamlConfiguration;
